+++ src/PluginValidate/Finder/TablePrefixFinder.php
@@ -24,8 +24,19 @@ class TablePrefixFinder extends TableFinder
 
     public function findTokens($file, FileTokens $fileTokens)
     {
-        foreach ($this->findTables($file) as $table) {
-            $fileTokens->compareStart($table);
+        $tables = $this->findTables($file);
+        $total  = count($tables);
+        for ($i = 0; $i < $total; ++$i) {
+            $fileTokens->compareStart($tables[$i]);
+
+            // This runs after every table except for the last one.
+            if ($i !== $total - 1) {
+                if (!$fileTokens->hasFoundAllTokens()) {
+                    break; // Found an invalid table name, can stop.
+                }
+                // Current table name valid, reset tokens so we can see if the next table is valid or not.
+                $fileTokens->resetTokens();
+            }
         }
     }
 }