+++ tests/PluginValidate/Finder/TablePrefixFinderTest.php
@@ -1,0 +1,41 @@
+<?php
+
+/*
+ * This file is part of the Moodle Plugin CI package.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * Copyright (c) 2017 Blackboard Inc. (http://www.blackboard.com)
+ * License http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace Moodlerooms\MoodlePluginCI\Tests\PluginValidate\Finder;
+
+use Moodlerooms\MoodlePluginCI\PluginValidate\Finder\FileTokens;
+use Moodlerooms\MoodlePluginCI\PluginValidate\Finder\TablePrefixFinder;
+
+class TablePrefixFinderTest extends \PHPUnit_Framework_TestCase
+{
+    public function testFindTokens()
+    {
+        $file       = __DIR__.'/../../Fixture/moodle-local_travis/db/install.xml';
+        $fileTokens = FileTokens::create('db/install.xml')->mustHave('local_travis');
+
+        $finder = new TablePrefixFinder();
+        $finder->findTokens($file, $fileTokens);
+
+        $this->assertTrue($fileTokens->hasFoundAllTokens());
+    }
+
+    public function testFindTokensFail()
+    {
+        $file       = __DIR__.'/../../Fixture/bad-install.xml';
+        $fileTokens = FileTokens::create('db/install.xml')->mustHave('local_travis');
+
+        $finder = new TablePrefixFinder();
+        $finder->findTokens($file, $fileTokens);
+
+        $this->assertFalse($fileTokens->hasFoundAllTokens());
+    }
+}